Update zombiedice.java
made changes to the while loop at the bottom, it now has if and else if statements for when brains or shotgun limits are reached. still need to add feature to allow other players to continue when a player gets 3 shotguns or chooses to cash out as well as keep score.